add support for dynamic plan item injection into case model
