+++ modules/flowable-cmmn-engine/src/main/java/org/flowable/cmmn/engine/impl/behavior/impl/StageActivityBehavior.java
@@ -38,7 +38,7 @@ public class StageActivityBehavior extends CoreCmmnTriggerableActivityBehavior i
     
     @Override
     public void execute(CommandContext commandContext, PlanItemInstanceEntity planItemInstanceEntity) {
-        if (planItemInstanceEntity.getPlanItem().getName() != null) {
+        if (planItemInstanceEntity.getName() == null && planItemInstanceEntity.getPlanItem().getName() != null) {
             Expression nameExpression = CommandContextUtil.getExpressionManager(commandContext).createExpression(planItemInstanceEntity.getPlanItem().getName());
             planItemInstanceEntity.setName(nameExpression.getValue(planItemInstanceEntity).toString());
         }
