+++ modules/flowable-cmmn-engine/src/main/java/org/flowable/cmmn/engine/impl/persistence/entity/PlanItemInstanceEntityBuilder.java
@@ -21,7 +21,7 @@ import org.flowable.cmmn.model.PlanItem;
  *
  * @author Micha Kiener
  */
-public interface PlanItemInstanceBuilder {
+public interface PlanItemInstanceEntityBuilder {
 
     /**
      * Set the plan item for the new instance to be based on. This is a mandatory information to be set.
@@ -29,7 +29,16 @@ public interface PlanItemInstanceBuilder {
      * @param planItem the plan item to base the new instance on
      * @return the builder instance for method chaining
      */
-    PlanItemInstanceBuilder planItem(PlanItem planItem);
+    PlanItemInstanceEntityBuilder planItem(PlanItem planItem);
+
+    /**
+     * Optionally override the name for this plan item instance and don't create it based on its plan item model. If set, it has priority over the
+     * default given by its plan item model.
+     *
+     * @param name the optional name to be used (overridden) for this plan item instance
+     * @return the builder instance for method chaining
+     */
+    PlanItemInstanceEntityBuilder name(String name);
 
     /**
      * Set the id of the case definition this plan item instance is part of. This is a mandatory information to be set.
@@ -37,42 +46,42 @@ public interface PlanItemInstanceBuilder {
      * @param caseDefinitionId the id of the case definition the plan item is a part of
      * @return the builder instance for method chaining
      */
-    PlanItemInstanceBuilder caseDefinitionId(String caseDefinitionId);
+    PlanItemInstanceEntityBuilder caseDefinitionId(String caseDefinitionId);
 
     /**
      * Set the id of the case instance the plan item instance is a direct or indirect child of. This is a mandatory information to be set.
      * @param caseInstanceId the id of the case instance the new plan item instance is a part of
      * @return the builder instance for method chaining
      */
-    PlanItemInstanceBuilder caseInstanceId(String caseInstanceId);
+    PlanItemInstanceEntityBuilder caseInstanceId(String caseInstanceId);
 
     /**
      * Set the id of the stage plan item instace the new plan item instance is a direct child of.
      * @param stagePlanItemInstanceId the id of the parent stage instance id for the new plan item instance
      * @return the builder instance for method chaining
      */
-    PlanItemInstanceBuilder stagePlanItemInstanceId(String stagePlanItemInstanceId);
+    PlanItemInstanceEntityBuilder stagePlanItemInstanceId(String stagePlanItemInstanceId);
 
     /**
      * Set the id of the tenant for the new plan item instance.
      * @param tenantId the id of the tenant the new plan item instance belongs to
      * @return the builder instance for method chaining
      */
-    PlanItemInstanceBuilder tenantId(String tenantId);
+    PlanItemInstanceEntityBuilder tenantId(String tenantId);
 
     /**
      * Optionally add any variables to be set on the new plan item instance as local variables.
      * @param localVariables an optional map of variables to set as local values for the new plan item instance
      * @return the builder instance for method chaining
      */
-    PlanItemInstanceBuilder localVariables(Map<String, Object> localVariables);
+    PlanItemInstanceEntityBuilder localVariables(Map<String, Object> localVariables);
 
     /**
      * Set true, if the new plan item instance to be created should be added to its parent, false otherwise.
      * @param addToParent true, if the plan item instance should be added to its parent
      * @return the builder instance for method chaining
      */
-    PlanItemInstanceBuilder addToParent(boolean addToParent);
+    PlanItemInstanceEntityBuilder addToParent(boolean addToParent);
 
     /**
      * Invoke this method to suppress any exceptions thrown when evaluating the plan item name expression. This might be necessary, if not all of the necessary
@@ -81,7 +90,7 @@ public interface PlanItemInstanceBuilder {
      * @param silentNameExpressionEvaluation true, if the name expression evaluation should ignore any exception thrown
      * @return the builder instance for method chaining
      */
-    PlanItemInstanceBuilder silentNameExpressionEvaluation(boolean silentNameExpressionEvaluation);
+    PlanItemInstanceEntityBuilder silentNameExpressionEvaluation(boolean silentNameExpressionEvaluation);
 
     /**
      * Checks for all necessary values to be present within the builder, creates a new plan item instance and returns it.
