+++ modules/flowable-cmmn-engine/src/test/resources/org/flowable/cmmn/test/dynamic/ChangeStateActivateHumanTaskTest.testActivateHumanTask.cmmn
@@ -29,8 +29,8 @@
                 </planItemOnPart>
             </sentry>
 
-            <humanTask id="task1" name="Task 1" flowable:assignee="kermit" />
-            <humanTask id="task2" name="Task 2" flowable:candidateGroups="test" />
+            <humanTask id="task1" name="Task One" flowable:assignee="kermit" />
+            <humanTask id="task2" name="Task Two" flowable:candidateGroups="test" />
 
         </casePlanModel>
     </case>
