+++ modules/flowable-cmmn-engine/src/test/resources/org/flowable/cmmn/test/dynamic/ChangeStateTest.testChangeHumanTaskInStage.cmmn
@@ -23,8 +23,8 @@
                         <standardEvent>complete</standardEvent>
                     </planItemOnPart>
                 </sentry>
-                <humanTask id="task1" name="Task 1" flowable:assignee="kermit" />
-                <humanTask id="task2" name="Task 2" flowable:candidateGroups="test" />
+                <humanTask id="task1" name="Task One" flowable:assignee="kermit" />
+                <humanTask id="task2" name="Task Two" flowable:candidateGroups="test" />
             </stage>
 
         </casePlanModel>
