+++ modules/flowable-cmmn-engine/src/test/resources/org/flowable/cmmn/test/jupiter/FlowableCmmnJupiterTest.extensionUsageExample.cmmn
@@ -3,7 +3,7 @@
     targetNamespace="http://flowable.org/cmmn">
     <case id="extensionUsage">
         <casePlanModel id="myPlanModel" name="My CasePlanModel">
-            <planItem id="planItem1" name="The task" definitionRef="theTask" />
+            <planItem id="planItem1" name="The Task" definitionRef="theTask" />
             <humanTask id="theTask" name="The Task" isBlocking="true" />
         </casePlanModel>
     </case>
