+++ modules/flowable-cmmn-engine/src/test/resources/org/flowable/cmmn/test/one-task-model.cmmn
@@ -7,7 +7,7 @@
     targetNamespace="http://flowable.org/cmmn">
     <case id="oneTaskCase">
         <casePlanModel id="myPlanModel" name="My CasePlanModel">
-            <planItem id="planItem1" name="The task" definitionRef="theTask" />
+            <planItem id="planItem1" name="The Task" definitionRef="theTask" />
             <task id="theTask" name="The Task" isBlocking="true" />
         </casePlanModel>
     </case>
