+++ modules/flowable-cmmn-engine/src/test/resources/org/flowable/cmmn/test/task/CmmnTaskServiceTest.testOneHumanTaskExpressionCase.cmmn
@@ -9,7 +9,7 @@
 
     <case id="oneHumanTaskCase">
         <casePlanModel id="myPlanModel" name="My CasePlanModel">
-            <planItem id="planItem1" name="The Task" definitionRef="theTask" />
+            <planItem id="planItem1" name="The Task ${var1}" definitionRef="theTask" />
             <humanTask id="theTask" name="The Task ${var1}" isBlocking="true" flowable:assignee="johnDoe">
                 <documentation>This is a test ${var2}</documentation>
             </humanTask>
